requirements.txt
kilter.service >= 0.3
config >= 0.5.0
checkdmarc ~= 5.3
